kafka-api/requirements.txt
fastapi
uvicorn
kafka-python
pandas
